refactor: reorganize API imports and rename login/logout files for clarity
