+++ src/api/mutations/index.ts
@@ -2,8 +2,6 @@ export * from './useAcceptFriendsRequests';
 export * from './useBlockFriend';
 export * from './useCreateFriendsRequests';
 export * from './useDeleteAvatar';
-export * from './useLogin';
-export * from './useLogout';
 export * from './useMailVerification';
 export * from './useOAuthLogin';
 export * from './useRegister';
