+++ src/app/auth/(include-footer)/login-email/page.tsx
@@ -2,11 +2,11 @@
 
 import { type FormEvent, useState } from 'react';
 import { toast } from 'sonner';
-import { useLogin } from '~/api';
 import { extractErrorData } from '~/api/base';
 import { MenuSelector, MenuSelectorBack } from '~/components/ui';
 import { env } from '~/constants/variables';
 import { InputForm } from '../../_components/input-form';
+import { useLogin } from './useLogin';
 
 const Page = () => {
   const [email, setEmail] = useState<string>('');
